feat: cohort table, CSV export, PDF generator, reports config (#238 #239 #240 #241)
