+++ analytics/lib/pdf-generator.ts
@@ -1,0 +1,2 @@
+export type { ReportSection } from './report';
+export { generatePdfReport, getDashboardDateRange } from './report';
